Refactor(src/modules/modalities/repositories) Add findById Method Rename folder name  src/modules/modalities/useCases/ListModalities/ to src/modules/modalities/useCases/ListModalities/
PLAYT38
