+++ src/modules/modalities/repositories/IModalitiesRepository.ts
@@ -5,6 +5,7 @@ interface IModalitiesRepository {
   create(data: ICreateModalityDTO): Promise<void>;
   list(): Promise<Modalities[]>;
   findByName(name: string): Promise<Modalities>;
+  findById(id: string): Promise<Modalities>;
   update(data: ICreateModalityDTO): Promise<Modalities>;
   delete(id: string): Promise<void>;
 }
