+++ src/routes/modalities.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 
 import { CreateModalityControler } from "../modules/modalities/useCases/createModality/CreateModalityController";
-import { ListModalityController } from "../modules/modalities/useCases/ListModalities/ListModalityController";
+import { ListModalityController } from "../modules/modalities/useCases/listModalities/ListModalityController";
 import { UpdateModalityController } from "../modules/modalities/useCases/updateModality/UpdateModalityController";
 
 const modalitiesRoutes = Router();
